apps/smart-writer/pyproject.toml
[project]
name = "smart-writer"
version = "0.1.0"
description = "Multi-agent writing pipeline (PydanticAI + LangGraph); fresh fork for new experiments"
requires-python = ">=3.12,<3.13"

dependencies = [
  "requests>=2.31,<3",
  "httpx>=0.27,<0.29",
  "trafilatura>=1.8,<2",
  "python-dotenv>=1,<2",

  # Keep your deliberate pydantic ceiling if you know 2.12 breaks your stack:
  "pydantic>=2.11.9,<2.12",

  # pydantic-ai is major 1 now; cap at next major:
  "pydantic-ai-slim[openai]>=1.59,<2",


  "langgraph>=1.0.6,<1.1",

  "supabase>=2.28,<3",

  # If you’re okay moving with logfire’s current major:
  "logfire>=4.24,<5",

  "fastapi>=0.115,<1",
  "uvicorn[standard]>=0.32,<1",
]

# Dev tooling: use [dependency-groups] (not optional-dependencies) so `uv sync` installs
# the default `dev` group — see https://docs.astral.sh/uv/concepts/projects/dependency-groups/
[dependency-groups]
dev = [
  "ruff>=0.8",
  "mypy>=1.13",
  "bandit[toml]>=1.7",
  "pytest>=8",
  "pytest-asyncio>=0.24,<1",
  "pytest-cov>=4",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

[tool.uv]
managed = true
package = false
# Install dependency-groups.dev by default (uv default-groups defaults to ["dev"]).
